CRY-4 - fix in base ai commenti della pull request(aggiornato joi)

+++ priv/migrations/20190704081945_products_price_nullable.js
@@ -1,7 +1,6 @@
-
 exports.up = (knex, Promise) =>
   knex.schema.alterTable('products', (t) => {
-    t.decimal('price').alter() // PRECISION=8 SCALE=2
+    t.integer('price').alter()
   })
 
 exports.down = (knex, Promise) => {
